pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "gpa"
version = "0.0.1"
description = "Open Graphics Profiler for Agents - live graphics debugger"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "PyYAML>=6.0",
    "anthropic>=0.40",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "httpx>=0.27",
]

[project.scripts]
gpa-curate = "gpa.eval.curation.pipeline:main"
